src/main/java/com/cryptoai/javaapi/binanceconnection/web/exception/RestExceptionHandler.java
package com.cryptoai.javaapi.binanceconnection.web.exception;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class RestExceptionHandler {

    @ExceptionHandler(WrongDateFormatException.class)
    public ResponseEntity<ErrorResponse> handleWrongDateFormatException(Exception exception){

        ErrorResponse error = new ErrorResponse(
                HttpStatus.BAD_REQUEST.value(),
                exception.getMessage(),
                System.currentTimeMillis()
        );

        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
    }

    @ExceptionHandler(NonexistentZipFileException.class)
    public ResponseEntity<ErrorResponse> handleNonexistentZipFileException(Exception exception){

        ErrorResponse error = new ErrorResponse(
                HttpStatus.NOT_FOUND.value(),
                exception.getMessage(),
                System.currentTimeMillis()
        );

        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
    }
}
